Add GradingScheme interface to Certificate.ts and refactor CourseForm and CoursesTable components
