+++ src/app/(admin)/admin/certificates/Certificate.ts
@@ -1,6 +1,12 @@
 import { Resource } from '@/app/(admin)/admin-core/repository/repository';
 
+export interface GradingScheme {
+  level: number;
+  grade: string;
+}
+
 export interface Certificate extends Resource {
   name: string;
   courses: string[];
+  gradingScheme: GradingScheme[];
 }
